[Backport] Security bug 905509 (1/13)
Manual backport of original patch by Antoine Labour
<piman@chromium.org>:
Fix ImplementationBase::SetBucketContents to take a uint32_t

The command takes a uint32_t, so any size_t would be silently truncated.
Clarify the proper range at the source.

Bug: 905509
Reviewed-on: https://chromium-review.googlesource.com/c/1336913

+++ chromium/gpu/command_buffer/client/gles2_implementation.cc
@@ -3481,7 +3481,7 @@ void GLES2Implementation::GetActiveUniformBlockiv(
 bool GLES2Implementation::GetActiveUniformsivHelper(
     GLuint program, GLsizei count, const GLuint* indices,
     GLenum pname, GLint* params) {
-  base::CheckedNumeric<size_t> bytes = static_cast<size_t>(count);
+  base::CheckedNumeric<uint32_t> bytes = count;
   bytes *= sizeof(GLuint);
   if (!bytes.IsValid()) {
     SetGLError(GL_INVALID_VALUE, "glGetActiveUniformsiv", "count overflow");
